+++ src/authentication/authenticationState.ts
@@ -1,6 +1,4 @@
 import { z } from "zod";
-import { authenticationStore } from "./LoginRedirect";
-import { AuthenticationContextType } from "./AuthenticationContext";
 
 const merchantInfoSchema = z.object({
   organization_name: z.string(),
@@ -8,6 +6,8 @@ const merchantInfoSchema = z.object({
   email: z.string(),
   token: z.string()
 });
+
+export type MerchantInfo = z.infer<typeof merchantInfoSchema>;
 
 const currentSessionSchema = z
   .discriminatedUnion("type", [
@@ -68,136 +68,6 @@ export const empty: AuthenticationState = {
   currentSession: null
 };
 
-const NONCE_DURATION = 15 * 60 * 1000;
-
-export function setUserNonceByState(state: string, nonce: string) {
-  const data = authenticationStore.get();
-  authenticationStore.set({
-    ...data,
-    userNonceByState: {
-      ...data.userNonceByState,
-      [state]: { nonce, exp: Date.now() + NONCE_DURATION }
-    }
-  });
-}
-
-export function getUserNonceByState(state: string): string | undefined {
-  const data = authenticationStore.get();
-  return data.userNonceByState[state]?.nonce;
-}
-
-export function deleteUserNonceByState(state: string) {
-  const data = authenticationStore.get();
-  const { [state]: deleted, ...userNonceByState } = data.userNonceByState;
-  authenticationStore.set({ ...data, userNonceByState });
-}
-
-export function setAdminNonceByState(state: string, nonce: string) {
-  const data = authenticationStore.get();
-  authenticationStore.set({
-    ...data,
-    adminNonceByState: {
-      ...data.adminNonceByState,
-      [state]: { nonce, exp: Date.now() + NONCE_DURATION }
-    }
-  });
-}
-
-export function getAdminNonceByState(state: string): string | undefined {
-  const data = authenticationStore.get();
-  return data.adminNonceByState[state]?.nonce;
-}
-
-export function deleteAdminNonceByState(state: string) {
-  const data = authenticationStore.get();
-  const { [state]: deleted, ...adminNonceByState } = data.adminNonceByState;
-  authenticationStore.set({ ...data, adminNonceByState });
-}
-
-export function setUserSession(fiscal_code: string, session: UserSession) {
-  const data = authenticationStore.get();
-  authenticationStore.set({
-    ...data,
-    userSessionByFiscalCode: {
-      ...data.userSessionByFiscalCode,
-      [fiscal_code]: session
-    }
-  });
-}
-
-export function setAdminSession(name: string, session: AdminSession) {
-  const data = authenticationStore.get();
-  authenticationStore.set({
-    ...data,
-    adminSessionByName: { ...data.adminSessionByName, [name]: session }
-  });
-}
-
-export function setCurrentSession(session: CurrentSession) {
-  const data = authenticationStore.get();
-  authenticationStore.set({
-    ...data,
-    currentSession: session
-  });
-}
-
-export function deleteSession(session: CurrentSession) {
-  const data = authenticationStore.get();
-  if (session?.type === "user") {
-    const { [session.userFiscalCode]: deleted, ...userSessionByFiscalCode } =
-      data.userSessionByFiscalCode;
-    authenticationStore.set({
-      ...data,
-      userSessionByFiscalCode
-    });
-  }
-  if (session?.type === "admin") {
-    const { [session.name]: deleted, ...adminSessionByName } =
-      data.adminSessionByName;
-    authenticationStore.set({
-      ...data,
-      adminSessionByName
-    });
-  }
-}
-
-export function getUserToken(): string {
-  const data = authenticationStore.get();
-  if (data.currentSession?.type === "user") {
-    return (
-      data.userSessionByFiscalCode[data.currentSession.userFiscalCode]?.token ??
-      ""
-    );
-  }
-  return "";
-  // returning empty, invalid or expired token here is fine since authentication errors are handled on401
-}
-
-export function getMerchantToken(): string {
-  const data = authenticationStore.get();
-  if (data.currentSession?.type === "user") {
-    const merchantFiscalCode = data.currentSession.merchantFiscalCode;
-    return (
-      data.userSessionByFiscalCode[
-        data.currentSession.userFiscalCode
-      ]?.merchants.find(
-        merchant => merchant.organization_fiscal_code === merchantFiscalCode
-      )?.token ?? ""
-    );
-  }
-  return "";
-  // returning empty, invalid or expired token here is fine since authentication errors are handled on401
-}
-
-export function getAdminToken(): string {
-  const data = authenticationStore.get();
-  if (data.currentSession?.type === "admin") {
-    return data.adminSessionByName[data.currentSession.name]?.token ?? "";
-  }
-  return "";
-  // returning empty, invalid or expired token here is fine since authentication errors are handled on401
-}
-
 export function excludeExpiredTokens(
   data: AuthenticationState
 ): AuthenticationState {
@@ -219,39 +89,4 @@ export function excludeExpiredTokens(
       Object.entries(data.adminSessionByName).filter(([, { exp }]) => exp > now)
     )
   };
-}
-
-export function getCurrentUserFiscalCode(state: AuthenticationContextType) {
-  if (state.currentSession && state.currentSession.type === "user") {
-    return state.currentSession.userFiscalCode;
-  }
-}
-
-export function getCurrentUserSession(state: AuthenticationContextType) {
-  if (state.currentSession && state.currentSession.type === "user") {
-    return state.userSessionByFiscalCode[state.currentSession.userFiscalCode];
-  }
-}
-
-export function getCurrentMerchantFiscalCode(state: AuthenticationContextType) {
-  if (state.currentSession && state.currentSession.type === "user") {
-    return state.currentSession.merchantFiscalCode;
-  }
-}
-
-export function getCurrentMerchant(state: AuthenticationContextType) {
-  if (state.currentSession && state.currentSession.type === "user") {
-    const merchantFiscalCode = state.currentSession.merchantFiscalCode;
-    return state.userSessionByFiscalCode[
-      state.currentSession.userFiscalCode
-    ]?.merchants.find(
-      merchant => merchant.organization_fiscal_code === merchantFiscalCode
-    );
-  }
-}
-
-export function getCurrentAdminSession(state: AuthenticationContextType) {
-  if (state.currentSession && state.currentSession.type === "admin") {
-    return state.adminSessionByName[state.currentSession.name];
-  }
 }
