+++ src/components/Requests/RequestsDetails.tsx
@@ -5,7 +5,6 @@ import { useTooltip, Severity } from "../../context/tooltip";
 import { Agreement, EntityType } from "../../api/generated_backoffice";
 import { getEntityTypeLabel } from "../../utils/strings";
 import { useAuthentication } from "../../authentication/AuthenticationContext";
-import { getCurrentUserSession } from "../../authentication/authenticationHelpers";
 import RequestItem from "./RequestsDetailsItem";
 import RequestsDocuments from "./RequestsDocuments";
 import AssignRequest from "./AssignRequest";
@@ -25,7 +24,7 @@ const RequestsDetails = ({
   const { triggerTooltip } = useTooltip();
 
   const authentication = useAuthentication();
-  const user = getCurrentUserSession(authentication);
+  const user = authentication.currentUserSession;
 
   const entityType = original.entityType as EntityType;
