+++ src/pages/Dashboard.tsx
@@ -10,14 +10,13 @@ import Profile from "../components/Profile/Profile";
 import ProfileData from "../components/ProfileData/ProfileData";
 import { RootState } from "../store/store";
 import { useAuthentication } from "../authentication/AuthenticationContext";
-import { getCurrentUserSession } from "../authentication/authenticationHelpers";
 
 const Dashboard = () => {
   const [tab, setTab] = useState(0);
   const agreement = useSelector((state: RootState) => state.agreement.value);
 
   const authentication = useAuthentication();
-  const user = getCurrentUserSession(authentication);
+  const user = authentication.currentUserSession;
 
   function hasStateSection(state: any) {
     return (
